requirements.txt
asgiref==3.7.2
Django==4.2.3
django-stdimage==6.0.1
Pillow==10.0.0
sqlparse==0.4.4
